Fixed integer comparison warning.
Github-Pull: #7023
Rebased-From: 4d29032a6437eaa147a69ce2857fb243bf3a1e49

+++ src/main.cpp
@@ -2381,7 +2381,7 @@ bool FindBlockPos(CValidationState &state, CDiskBlockPos &pos, unsigned int nAdd
         pos.nPos = vinfoBlockFile[nFile].nSize;
     }
 
-    if (nFile != nLastBlockFile) {
+    if ((int)nFile != nLastBlockFile) {
         if (!fKnown) {
             LogPrintf("Leaving block file %i: %s\n", nFile, vinfoBlockFile[nFile].ToString());
         }
